requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
scikit-learn>=1.0.0
jupyter>=1.0.0
tensorboard>=2.7.0
